Use a custom /etc/rc script on first boot instead of rc.sysmerge script.
This is used to set up the VM before the first boot, randomize kernel
and inodes, and eventually run syspatch (on -stable).

This is another experimental approach as an alternative to the
previous commit.  The simplified rc.sysmerge lacks a few things, like
randomization of inode generation numbers (fsirand).  The third
approach will most likely be a specialized auto-installer and we'll
see which one is the best.